Вынес onFetchError в main
